Move multPlusTransA into Householder_DDRB; thin TridiagonalDecomposition getQ (Phase 1.2c)
- multPlusTransA (concurrent C += A^T*B for row block vectors) -> Householder_DDRB (L3),
  beside multTransA_vecCol; qualified its blockMultPlusTransA call to match the class style.
  TridiagonalDecompositionHouseholder_DDRB.decompose now calls Householder_DDRB.multPlusTransA;
  added //CONCURRENT_MACRO Householder_DDRB Householder_MT_DDRB so the _MT decompose keeps the
  concurrent variant; dropped the now-unused InnerMultiplication_DDRB static import.
  (Note: the unrelated generic MatrixMult_DDRB.multPlusTransA used in getQ is untouched.)
- getQ: replaced the inline last-reflector-block computation with MatrixOps_DDRB.lastBlockStart(N, blockLength).
- Removed the stray commented System.out.println in decompose().
- Tests mirror the homes: multPlusTransA serial test moved into TestHouseholder_DDRB (+ a matching
  serial-vs-_MT compare in TestHouseholder_MT_DDRB); removed from TestTridiagonalDecompositionHouseholder_DDRB
  (dropped its orphaned DSubmatrixD1 import).

TridiagonalDecompositionHouseholder_DDRB now keeps only its unique code (getT, getDiagonal,
copyZeros/replaceZeros [until Phase 3], and the tridiagonal trailing update).
Serial + _MT block decomposition/solver tests green (ddense + fdense).
